requirements.txt
mss
pillow
pytesseract
soundcard
PyAudio
openai-whisper
google-generativeai
PyQt5
pywin32
ollama 